6.0 add test
