+++ hibernate-core/src/test/java/org/hibernate/orm/test/event/EmbeddableCallbackTest.java
@@ -4,43 +4,49 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.event;
+package org.hibernate.orm.test.event;
 
 import javax.persistence.Embeddable;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.PrePersist;
 
-import org.hibernate.jpa.test.BaseEntityManagerFunctionalTestCase;
+import org.hibernate.boot.MetadataSources;
+import org.hibernate.orm.test.SessionFactoryBasedFunctionalTest;
 
 import org.hibernate.testing.TestForIssue;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.hibernate.testing.transaction.TransactionUtil.doInJPA;
 import static org.junit.Assert.assertEquals;
 
 /**
  * @author Vlad Mihalcea
  */
 @TestForIssue(jiraKey = "HHH-12326")
-public class EmbeddableCallbackTest extends BaseEntityManagerFunctionalTestCase {
+public class EmbeddableCallbackTest extends SessionFactoryBasedFunctionalTest {
 
 	@Override
-	protected Class<?>[] getAnnotatedClasses() {
-		return new Class<?>[] { Employee.class };
+	protected void applyMetadataSources(MetadataSources metadataSources) {
+		super.applyMetadataSources( metadataSources );
+		metadataSources.addAnnotatedClass( Employee.class );
+	}
+
+	@Override
+	protected boolean exportSchema() {
+		return true;
 	}
 
 	@Test
 	public void test() {
-		doInJPA( this::entityManagerFactory, entityManager -> {
+		sessionFactoryScope().inTransaction( session -> {
 			Employee employee = new Employee();
 			employee.id = 1;
 
-			entityManager.persist( employee );
+			session.persist( employee );
 		} );
 
-		doInJPA( this::entityManagerFactory, entityManager -> {
-			Employee employee = entityManager.find( Employee.class, 1 );
+		sessionFactoryScope().inTransaction( session -> {
+			Employee employee = session.find( Employee.class, 1 );
 
 			assertEquals( "Vlad", employee.name );
 			assertEquals( "Developer Advocate", employee.details.jobTitle );
