+++ hibernate-core/src/test/java/org/hibernate/orm/test/event/PreInsertEventListenerVetoUnidirectionalTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.event;
+package org.hibernate.orm.test.event;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
@@ -13,20 +13,21 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToOne;
 
-import org.hibernate.TransientObjectException;
 import org.hibernate.action.internal.EntityActionVetoException;
+import org.hibernate.boot.MetadataSources;
 import org.hibernate.event.service.spi.EventListenerRegistry;
 import org.hibernate.event.spi.EventType;
 import org.hibernate.event.spi.PreInsertEventListener;
+import org.hibernate.orm.test.SessionFactoryBasedFunctionalTest;
 
 import org.hibernate.testing.DialectChecks;
 import org.hibernate.testing.RequiresDialectFeature;
 import org.hibernate.testing.TestForIssue;
-import org.hibernate.testing.junit4.BaseCoreFunctionalTestCase;
-import org.junit.Test;
+import org.hibernate.testing.junit5.ExpectedException;
+import org.hibernate.testing.junit5.FailureExpected;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Test;
 
-import static org.hibernate.testing.transaction.TransactionUtil.doInHibernate;
-import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 /**
@@ -34,15 +35,17 @@ import static org.junit.Assert.fail;
  */
 @RequiresDialectFeature(DialectChecks.SupportsIdentityColumns.class)
 @TestForIssue(jiraKey = "HHH-11721")
-public class PreInsertEventListenerVetoBidirectionalTest extends BaseCoreFunctionalTestCase {
+@FailureExpected( value= "fetching database snapshot not yet implemented" )
+public class PreInsertEventListenerVetoUnidirectionalTest extends SessionFactoryBasedFunctionalTest {
 	@Override
-	protected Class<?>[] getAnnotatedClasses() {
-		return new Class<?>[] { Child.class, Parent.class };
+	protected void applyMetadataSources(MetadataSources metadataSources) {
+		super.applyMetadataSources( metadataSources );
+		metadataSources.addAnnotatedClass( Child.class );
+		metadataSources.addAnnotatedClass( Parent.class );
 	}
 
-	@Override
+	@AfterAll
 	protected void afterSessionFactoryBuilt() {
-		super.afterSessionFactoryBuilt();
 		EventListenerRegistry registry = sessionFactory().getServiceRegistry()
 				.getService( EventListenerRegistry.class );
 		registry.appendListeners(
@@ -51,17 +54,18 @@ public class PreInsertEventListenerVetoBidirectionalTest extends BaseCoreFunctio
 		);
 	}
 
-	@Test(expected = EntityActionVetoException.class)
+	@Test
+	@ExpectedException(value = EntityActionVetoException.class)
 	public void testVeto() {
-		doInHibernate( this::sessionFactory, session -> {
+		sessionFactoryScope().inTransaction( session -> {
 			Parent parent = new Parent();
 			parent.setField1( "f1" );
 			parent.setfield2( "f2" );
 
 			Child child = new Child();
-			parent.setChild( child );
+			child.setParent( parent );
 
-			session.save( parent );
+			session.save( child );
 		} );
 
 		fail( "Should have thrown EntityActionVetoException!" );
@@ -69,12 +73,11 @@ public class PreInsertEventListenerVetoBidirectionalTest extends BaseCoreFunctio
 
 	@Entity(name = "Child")
 	public static class Child {
-
 		@Id
 		@GeneratedValue(strategy = GenerationType.IDENTITY)
 		private Integer id;
 
-		@OneToOne
+		@OneToOne(cascade = CascadeType.ALL)
 		private Parent parent;
 
 		public Integer getId() {
@@ -96,17 +99,11 @@ public class PreInsertEventListenerVetoBidirectionalTest extends BaseCoreFunctio
 
 	@Entity(name = "Parent")
 	public static class Parent {
-
 		@Id
 		@GeneratedValue(strategy = GenerationType.IDENTITY)
 		private Integer id;
-
 		private String field1;
-
 		private String field2;
-
-		@OneToOne(cascade = CascadeType.ALL, mappedBy = "parent")
-		private Child child;
 
 		public Integer getId() {
 			return id;
@@ -130,15 +127,6 @@ public class PreInsertEventListenerVetoBidirectionalTest extends BaseCoreFunctio
 
 		public void setfield2(String field2) {
 			this.field2 = field2;
-		}
-
-		public Child getChild() {
-			return child;
-		}
-
-		public void setChild(Child child) {
-			this.child = child;
-			child.setParent( this );
 		}
 	}
 }
